+++ tools/xdhq/wrappers/JRE/Event.java
@@ -23,4 +23,8 @@ public class Event {
 	public String
 		id,
 		action;
+		public Event() {
+			id = new String();
+			action = new String();
+		}
 };